+++ lib/document_auto_capture/conversion_extensions.dart
@@ -4,17 +4,31 @@ import 'package:dot_flutter_sdk_example/conversion_extensions.dart';
 extension DocumentAutoCaptureResultExt on DocumentAutoCaptureResult {
   Map<String, dynamic> toJson() {
     return {
-      'bgraRawImage': {
-        'bytes': bgraRawImage.bytes.toJson(),
-        'size': bgraRawImage.size.toJson(),
-      },
+      'bgraRawImage': bgraRawImage.toJson(),
+      'document': document?.toJson(),
       'content': content.toJson(),
-      'document': {
-        'confidence': document?.confidence,
-        'position': document?.position.toJson(),
-        'widthToHeightRatio': document?.widthToHeightRatio,
-        'imageParameters': document?.imageParameters.toJson(),
-      },
+    };
+  }
+}
+
+extension DocumentDetectorDocumentExt on DocumentDetectorDocument {
+  Map<String, dynamic> toJson() {
+    return {
+      'confidence': confidence,
+      'position': position.toJson(),
+      'widthToHeightRatio': widthToHeightRatio,
+      'imageParameters': imageParameters.toJson(),
+    };
+  }
+}
+
+extension DetectionPositionExt on DetectionPosition {
+  Map<String, dynamic> toJson() {
+    return {
+      'topLeft': topLeft.toJson(),
+      'topRight': topRight.toJson(),
+      'bottomRight': bottomRight.toJson(),
+      'bottomLeft': bottomLeft.toJson(),
     };
   }
 }
@@ -25,17 +39,6 @@ extension ImageParametersExt on ImageParameters {
       'brightness': brightness,
       'sharpness': sharpness,
       'hotspotsScore': hotspotsScore,
-    };
-  }
-}
-
-extension DetectionPositionExt on DetectionPosition {
-  Map<String, dynamic> toJson() {
-    return {
-      'bottomLeft': bottomLeft.toJson(),
-      'bottomRight': bottomRight.toJson(),
-      'topLeft': topLeft.toJson(),
-      'topRight': topRight.toJson(),
     };
   }
 }