+++ lib/magnifeye_liveness/conversion_extensions.dart
@@ -1,4 +1,4 @@
-import 'package:dot_face_lite/dot_face_lite.dart';
+import 'package:dot_face_core/dot_face_core.dart';
 import 'package:dot_flutter_sdk_example/conversion_extensions.dart';
 
 import '../face/conversion_extensions.dart';
@@ -6,15 +6,8 @@ import '../face/conversion_extensions.dart';
 extension MagnifEyeLivenessResultExt on MagnifEyeLivenessResult {
   Map<String, dynamic> toJson() {
     return {
-      'bgraRawImage': {
-        'bytes': bgraRawImage.bytes.toJson(),
-        'size': bgraRawImage.size.toJson(),
-      },
-      'face': {
-        'confidence': face.confidence,
-        'position': face.position.toJson(),
-        'imageParameters': face.imageParameters.toJson(),
-      },
+      'bgrRawImage': bgrRawImage.toJson(),
+      'face': face.toJson(),
       'content': content.toJson(),
     };
   }