+++ lib/magnifeye_liveness/magnifeye_liveness_result_screen.dart
@@ -1,11 +1,11 @@
 import 'dart:convert';
 
-import 'package:dot_face_lite/dot_face_lite.dart';
+import 'package:dot_face_core/dot_face_core.dart';
 import 'package:flutter/material.dart';
 
 import '../disposable_raw_image.dart';
 import '../json_encoder_factory.dart';
-import '../magnifeye_liveness/conversion_extensions.dart';
+import 'conversion_extensions.dart';
 
 class MagnifEyeLivenessResultScreen extends StatelessWidget {
   final MagnifEyeLivenessResult result;
@@ -55,7 +55,8 @@ class MagnifEyeLivenessResultScreen extends StatelessWidget {
   }
 
   Future<RawImage> getImageFromResult(MagnifEyeLivenessResult result) async {
-    final uiImage = await ImageFactory.create(result.bgraRawImage);
+    final uiImage =
+        await ImageFactory.createFromBgrRawImage(result.bgrRawImage);
     return DisposableRawImage(
       image: uiImage,
       color: Colors.black,